Like #13843 but based on top of `main`.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                                        │  before.txt   │             after.txt              │
                                                        │    sec/op     │   sec/op     vs base               │
FastRegexMatcher/#00-11                                    34.94n ±  1%   35.65n ± 0%   +2.05% (p=0.002 n=6)
FastRegexMatcher/foo-11                                    36.32n ±  0%   36.06n ± 0%   -0.70% (p=0.004 n=6)
FastRegexMatcher/^foo-11                                   34.24n ±  4%   33.78n ± 2%        ~ (p=0.169 n=6)
FastRegexMatcher/(foo|bar)-11                              47.35n ±  0%   47.14n ± 0%   -0.45% (p=0.002 n=6)
FastRegexMatcher/foo.*-11                                  83.58n ±  0%   83.66n ± 2%        ~ (p=0.145 n=6)
FastRegexMatcher/.*foo-11                                  99.32n ±  0%   99.33n ± 2%        ~ (p=0.310 n=6)
FastRegexMatcher/^.*foo$-11                                99.27n ±  0%   99.30n ± 0%        ~ (p=0.065 n=6)
FastRegexMatcher/^.+foo$-11                                99.60n ±  0%   99.61n ± 0%        ~ (p=0.450 n=6)
FastRegexMatcher/.?-11                                     55.55n ±  0%   55.56n ± 0%        ~ (p=0.314 n=6)
FastRegexMatcher/.*-11                                     66.83n ±  1%   68.41n ± 6%        ~ (p=0.071 n=6)
FastRegexMatcher/.+-11                                     70.67n ±  2%   71.02n ± 5%        ~ (p=0.420 n=6)
FastRegexMatcher/foo.+-11                                  83.88n ±  2%   83.92n ± 0%        ~ (p=0.818 n=6)
FastRegexMatcher/.+foo-11                                  99.61n ±  0%   99.60n ± 0%        ~ (p=0.909 n=6)
FastRegexMatcher/foo_.+-11                                 70.83n ±  0%   70.84n ± 1%        ~ (p=0.545 n=6)
FastRegexMatcher/foo_.*-11                                 70.84n ±  0%   70.84n ± 0%        ~ (p=0.470 n=6)
FastRegexMatcher/.*foo.*-11                                173.3n ±  1%   174.5n ± 7%        ~ (p=0.255 n=6)
FastRegexMatcher/.+foo.+-11                                183.9n ±  1%   183.9n ± 1%        ~ (p=0.779 n=6)
FastRegexMatcher/(?s:.*)-11                                34.35n ±  1%   34.06n ± 0%   -0.82% (p=0.002 n=6)
FastRegexMatcher/(?s:.+)-11                                35.09n ±  1%   34.74n ± 0%   -0.97% (p=0.002 n=6)
FastRegexMatcher/(?s:^.*foo$)-11                           97.00n ±  0%   96.98n ± 0%        ~ (p=0.972 n=6)
FastRegexMatcher/(?i:foo)-11                               63.74n ±  0%   63.47n ± 1%        ~ (p=0.394 n=6)
FastRegexMatcher/(?i:(foo|bar))-11                         150.0n ±  1%   142.8n ± 0%   -4.77% (p=0.002 n=6)
FastRegexMatcher/(?i:(foo1|foo2|bar))-11                   250.0n ±  1%   242.4n ± 0%   -3.00% (p=0.002 n=6)
FastRegexMatcher/^(?i:foo|oo)|(bar)$-11                    636.6n ±  0%   634.4n ± 1%        ~ (p=0.619 n=6)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-11       745.6n ±  4%   757.1n ± 6%        ~ (p=0.132 n=6)
FastRegexMatcher/((.*)(bar|b|buzz)(.+)|foo)$-11            441.1n ±  1%   437.5n ± 0%   -0.79% (p=0.009 n=6)
FastRegexMatcher/^$-11                                     34.54n ±  0%   35.27n ± 0%   +2.14% (p=0.002 n=6)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-11        162.6n ±  3%   166.1n ± 3%        ~ (p=0.818 n=6)
FastRegexMatcher/10\.0\.(1|2)\.+-11                        70.99n ±  0%   70.81n ± 0%   -0.25% (p=0.032 n=6)
FastRegexMatcher/10\.0\.(1|2).+-11                         70.86n ±  0%   70.83n ± 0%        ~ (p=0.450 n=6)
FastRegexMatcher/((fo(bar))|.+foo)-11                      189.3n ±  0%   180.9n ± 0%   -4.44% (p=0.002 n=6)
FastRegexMatcher/zQPbMkNO|NNSPdvMi|iWuuSoAl|qbvKM-11       190.2n ±  6%   194.0n ± 7%        ~ (p=0.310 n=6)
FastRegexMatcher/jyyfj00j0061|jyyfj00j0062|jyyfj9-11       182.1n ± 12%   186.8n ± 8%        ~ (p=0.394 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS-11       201.8n ±  0%   197.4n ± 1%   -2.16% (p=0.002 n=6)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-11       762.4n ±  3%   771.0n ± 2%        ~ (p=0.180 n=6)
FastRegexMatcher/(?i:(AAAAAAAAAAAAAAAAAAAAAAAA|BB-11       324.1n ±  0%   317.2n ± 0%   -2.14% (p=0.002 n=6)
FastRegexMatcher/(?i:(zQPbMkNO.*|NNSPdvMi.*|iWuuS#01-11   4703.0n ±  0%   989.6n ± 1%  -78.96% (p=0.002 n=6)
FastRegexMatcher/(?i:(.*zQPbMkNO|.*NNSPdvMi|.*iWu-11       6.150µ ±  0%   6.268µ ± 1%   +1.93% (p=0.002 n=6)
FastRegexMatcher/fo.?-11                                   83.54n ±  1%   83.29n ± 0%   -0.31% (p=0.043 n=6)
FastRegexMatcher/foo.?-11                                  82.87n ±  0%   82.86n ± 0%        ~ (p=0.574 n=6)
FastRegexMatcher/f.?o-11                                   71.48n ±  0%   71.45n ± 0%        ~ (p=0.937 n=6)
FastRegexMatcher/.*foo.?-11                                183.7n ±  0%   183.0n ± 4%        ~ (p=0.258 n=6)
FastRegexMatcher/.?foo.+-11                                173.9n ±  2%   173.1n ± 7%        ~ (p=0.368 n=6)
FastRegexMatcher/foo.?|bar-11                              133.8n ±  0%   129.5n ± 1%   -3.21% (p=0.002 n=6)
FastRegexMatcher/ſſs-11                                    36.05n ±  0%   36.09n ± 0%        ~ (p=0.193 n=6)
FastRegexMatcher/.*-.*-.*-.*-.*-11                         177.7n ±  2%   177.6n ± 2%        ~ (p=0.584 n=6)
FastRegexMatcher/(.+)-(.+)-(.+)-(.+)-(.+)-11               176.9n ±  2%   178.1n ± 0%   +0.68% (p=0.026 n=6)
FastRegexMatcher/((.*))(?i:f)((.*))o((.*))o((.*))-11       4.027µ ±  6%   4.062µ ± 0%        ~ (p=0.056 n=6)
FastRegexMatcher/((.*))f((.*))(?i:o)((.*))o((.*))-11       3.269µ ±  2%   3.291µ ± 1%        ~ (p=0.063 n=6)
geomean                                                    148.9n         144.0n        -3.25%
```